requirements.txt
azure-identity>=1.7.0
requests>=2.28.0
python-dateutil>=2.8.2
setuptools>=59.5.0